RDKTV-2340 RDKTV-2135 : Point to secure MTLS endpoints for FW Update
Reason for change:
The endpoints for xconf and ssr for Platco are currently pointed to non-secure TLS endpoints.
As a result attempts to reach out to the servers timeout leading to delays in Firmware version check (ENT-41001)
and firmware update (ENT-41005). Fixing it by pointing it to the correct secure MTLS endpoints.
Test Procedure: Refer to ticket
Risks: Medium

Signed-off-by: Anand Kandasamy <Anand_Kandasamy@comcast.com>
Change-Id: If2c6542dafd78a046a4a2e04a66ad5d070c20491